Struct lightning_invoice::SignedRawInvoice

pub struct SignedRawInvoice { /* private fields */ }

Represents a signed RawInvoice with cached hash. The signature is not checked and may be invalid.

Invariants

The hash has to be either from the deserialized invoice or from the serialized raw_invoice.

Implementations

impl SignedRawInvoice

pub fn into_parts(self) -> (RawInvoice, [u8; 32], InvoiceSignature)

Disassembles the SignedRawInvoice into its three parts:

raw invoice

hash of the raw invoice

signature

pub fn raw_invoice(&self) -> &RawInvoice

The RawInvoice which was signed.

pub fn signable_hash(&self) -> &[u8; 32]

The hash of the RawInvoice that was signed.

pub fn signature(&self) -> &InvoiceSignature

InvoiceSignature for the invoice.

pub fn recover_payee_pub_key(&self) -> Result<PayeePubKey, Error>

Recovers the public key used for signing the invoice from the recoverable signature.

pub fn check_signature(&self) -> bool

Checks if the signature is valid for the included payee public key or if none exists if it’s valid for the recovered signature (which should always be true?).