Function lightning_invoice::utils::create_invoice_from_channelmanager_with_description_hash_and_duration_since_epoch

pub fn create_invoice_from_channelmanager_with_description_hash_and_duration_since_epoch<M: Deref, T: Deref, ES: Deref, NS: Deref, SP: Deref, F: Deref, R: Deref, L: Deref>(
    channelmanager: &ChannelManager<M, T, ES, NS, SP, F, R, L>,
    node_signer: NS,
    logger: L,
    network: Currency,
    amt_msat: Option<u64>,
    description_hash: Sha256,
    duration_since_epoch: Duration,
    invoice_expiry_delta_secs: u32,
    min_final_cltv_expiry_delta: Option<u16>
) -> Result<Invoice, SignOrCreationError<()>>where
    M::Target: Watch<<SP::Target as SignerProvider>::Signer>,
    T::Target: BroadcasterInterface,
    ES::Target: EntropySource,
    NS::Target: NodeSigner,
    SP::Target: SignerProvider,
    F::Target: FeeEstimator,
    R::Target: Router,
    L::Target: Logger,

See create_invoice_from_channelmanager_with_description_hash This version can be used in a no_std environment, where std::time::SystemTime is not available and the current time is supplied by the caller.